pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = ""
authors = [{ name = "Võ Đình Đạt", email = "vodinhdat03@gmail.com" }]
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "gradio<5",
    "python-dotenv>=1.0.1,<2",
    "llama-index-embeddings-huggingface>=0.1.4,<0.2",
    "llama-index-llms-openai>=0.1.12,<0.2",
    "llama-index-vector-stores-chroma>=0.1.6,<0.2",
    "llama-index-llms-ollama>=0.1.2,<0.2",
    "llama-index-embeddings-ollama>=0.1.2,<0.2",
    "llama-index-readers-file>=0.1.11,<0.2",
    "llama-index>=0.10.22,<0.11",
    "transformers>=4.39.1,<5",
    "torch",
    "llama-index-callbacks-wandb>=0.1.2,<0.2",
    "llama-index-retrievers-bm25>=0.1.3,<0.2",
    "pymupdf>=1.24.3,<2",
    "tqdm>=4.66.4,<5",
    "requests>=2.32.3,<3",
    "pandas>=2.2.3,<3",
    "sentence-transformers>=3.2.0,<4",
    "pydantic==2.8.2",
]

[dependency-groups]
dev = [
    "ruff>=0.6.9,<0.7",
    "deptry>=0.20.0,<0.21",
    "pytest>=8.3.3,<9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.deptry.per_rule_ignores]
DEP002 = ["sentence-transformers"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".venv",
]